fixed issue when using file argument

+++ csv-html.py
@@ -2,6 +2,7 @@ import pandas as pd
 import argparse
 import xerox
 import os
+import tempfile
 from pandas.errors import ParserError
 
 parser = argparse.ArgumentParser(description='Convert Tab-delimited table to HTML table.')
@@ -9,7 +10,7 @@ parser.add_argument('--file',  type=str,
                     help='the path to the CSV file to convert')
 
 
-TEMP_FILENAME=os.path.join(os.getcwd(), 'csv_clipboard_temp.csv')
+TEMP_FILENAME=os.path.join(tempfile.gettempdir(), 'csv_clipboard_temp.csv')
 
 def convertToHtml(filepath):
     try:
@@ -17,7 +18,7 @@ def convertToHtml(filepath):
     except ParserError as pe:
         print(f"Invalid CSV format: {pe}")
         return
-    print(df.to_html(na_rep='', float_format='{0:.2f}'.format).replace(r'<td>nan</td>', "<td></td>"))
+    print(df.to_html(na_rep='', index=False, float_format='{0:.2f}'.format).replace(r'<td>nan</td>', "<td></td>"))
 
 def writeToTempFile(fileText):
     with open(TEMP_FILENAME, 'w') as f:
@@ -30,5 +31,6 @@ if __name__ == '__main__':
     if filepath is None:
         writeToTempFile(xerox.paste())
         filepath=TEMP_FILENAME
-    convertToHtml(TEMP_FILENAME)
-    os.remove(TEMP_FILENAME)
+    convertToHtml(filepath)
+    if os.path.isfile(TEMP_FILENAME):
+        os.remove(TEMP_FILENAME)
